pyproject.toml
[tool.poetry]
name = "thoughtgrove"
version = "0.1.0"
description = "A Digital Garden for Cultivating Ideas"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "thoughtgrove"}]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.110.0"
uvicorn = "^0.27.1"
pydantic = {extras = ["email"], version = "^2.6.1"}
motor = "^3.3.2"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.9"
openai = "^1.12.0"
python-dotenv = "^1.0.1"
pydantic-settings = "^2.8.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.23.5"
httpx = "^0.26.0"
black = "^24.2.0"
isort = "^5.13.2"
mypy = "^1.8.0"
flake8 = "^7.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py312']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3 